+++ plugins/module_utils/network/nxos/rm_templates/snmp_server.py
@@ -1487,6 +1487,7 @@ class Snmp_serverTemplate(NetworkTemplate):
                 $""", re.VERBOSE,
             ),
             "setval": _tmplt_users_auth,
+            "remval": "snmp-server user {{ user }}",
             "result": {
                 "users": {
                     "auth": [